+++ src/lib/eval.mli
@@ -12,11 +12,10 @@
 (** Evaluating and manipulating expressions *)
 
 exception Unknown_id of string
-exception Unbound_id of string
 exception Illegal_expr of Expr.t
 
 type env = (string * Expr.e_val) list
 
 val subst : (string * Expr.e_val) list -> Expr.t -> Expr.t
 
-val eval : (string * Expr.e_val option) list -> Expr.t -> Expr.e_val
+val eval : (string * Expr.e_val) list -> Expr.t -> Expr.e_val