+++ src/lib/simul.mli
@@ -19,19 +19,19 @@ type config = {
 
 val cfg : config
 
-type stimulus = Ident.t * Expr.e_val option  (** name, value (None for pure events) *)
-type response = Ident.t * Expr.e_val option   (** name, value (None for pure events) *)
+type stimulus = Ident.t * Expr.e_val (** name, value (Val_none for pure events) *)
+type response = Ident.t * Expr.e_val (** name, value (Val_none for pure events) *)
 
 type reaction = Types.date * string * Stimuli.stimuli list * response list * string
 
 type context = {  (** The simulator state *)
   c_date: Types.date;
-  c_inputs: (string * (Types.typ * Expr.e_val option)) list;   (** Global inputs *)
-  c_outputs: (string * (Types.typ * Expr.e_val option)) list;  (** Globals outputs *)
-  c_fns: (string * (Types.typ * Expr.e_val option)) list;      (* Global functions *)
-  c_vars: (string * (Types.typ * Expr.e_val option)) list;     (** Shared variables *)
-  c_evs: (string * (Types.typ * Expr.e_val option)) list;      (** Shared events *)
-  c_fsms: Fsm.inst list * Fsm.inst list;                       (** FSMs, partitioned into active and inactive subsets *)
+  c_inputs: (string * (Types.typ * Expr.e_val)) list;   (** Global inputs *)
+  c_outputs: (string * (Types.typ * Expr.e_val)) list;  (** Globals outputs *)
+  c_fns: (string * (Types.typ * Expr.e_val)) list;      (* Global functions *)
+  c_vars: (string * (Types.typ * Expr.e_val)) list;     (** Shared variables *)
+  c_evs: (string * (Types.typ * Expr.e_val)) list;      (** Shared events *)
+  c_fsms: Fsm.inst list * Fsm.inst list;                (** FSMs, partitioned into active and inactive subsets *)
   }
 
 exception OverReaction of Types.date
@@ -52,4 +52,4 @@ val run : Sysm.t -> context * (Types.date * response list) list
 (** {2 Printers} *)
 
 val dump_context : context -> unit
-val dump_reaction : int * (Ident.t * Expr.e_val option) list -> unit
+val dump_reaction : int * (Ident.t * Expr.e_val) list -> unit