+++ src/lib/stimuli.mli
@@ -11,17 +11,17 @@
 
 (** Events and stimuli *)
 
-type event = Types.date * Expr.e_val option
-  (** Single event at time [t]. Value is [None] for pure events *)
+type event = Types.date * Expr.e_val
+  (** Single event at time [t]. Value is [Val_none] for pure events *)
 
-type stimuli = Types.date * (Ident.t * Expr.e_val option) list
+type stimuli = Types.date * (Ident.t * Expr.e_val) list
   (** Set of named events at time [t] *)
 
-val mk_spor_event : Types.date list -> (Types.date * Expr.e_val option) list
-val mk_per_event : Types.date -> Types.date -> Types.date -> (Types.date * Expr.e_val option) list
-val mk_val_changes : (Types.date * Expr.e_val) list -> (Types.date * Expr.e_val option) list
+val mk_spor_event : Types.date list -> (Types.date * Expr.e_val) list
+val mk_per_event : Types.date -> Types.date -> Types.date -> (Types.date * Expr.e_val) list
+val mk_val_changes : (Types.date * Expr.e_val) list -> (Types.date * Expr.e_val) list
 
-val mk_stimuli : string -> Types.date * Expr.e_val option -> Types.date * (Ident.t * Expr.e_val option) list
+val mk_stimuli : string -> Types.date * Expr.e_val -> Types.date * (Ident.t * Expr.e_val) list
 
 val merge_stimuli : stimuli list list -> stimuli list