+++ src/lib/types.ml
@@ -224,6 +224,8 @@ let type_of_value = function
   | Expr.Val_bool _ -> TyBool
   | Expr.Val_enum c -> TyEnum [c]  (* TO FIX *)
   | Expr.Val_fn (args,body) -> TyArrow (TyProduct (List.map (function arg -> TyBool) args),TyBool) (* TO FIX ! *)
+  | Expr.Val_unknown -> new_type_var ()
+  | Expr.Val_none -> TyEvent
 
 (* Accessors *)